feat: add contextual hatch ribbon controls
