+++ packages/cad-viewer/src/component/common/MlHatchPatternDropdown.vue
@@ -1,59 +1,47 @@
 <template>
   <div class="ml-hatch-pattern-dropdown">
-    <el-dropdown
-      ref="dropdownRef"
+    <ElPopover
       v-model:visible="isOpen"
       trigger="click"
       placement="bottom-start"
       popper-class="ml-hatch-pattern-dropdown-popper"
       :disabled="disabled"
+      :offset="6"
+      :show-arrow="false"
+      persistent
     >
-      <div
-        class="ml-hatch-pattern-dropdown__trigger"
-        :class="{ 'is-disabled': disabled }"
-      >
-        <button
-          type="button"
-          class="ml-hatch-pattern-dropdown__main-button"
+      <template #reference>
+        <MlRibbonButton
+          :id="buttonId"
+          class="ml-hatch-pattern-dropdown__button"
+          :label="currentLabel"
           :disabled="disabled"
-          @click.stop="handleMainButtonClick"
+          :aria-label="accessibleLabel"
         >
-          <span class="ml-hatch-pattern-dropdown__value">
+          <template #icon>
             <span
               class="ml-hatch-pattern-dropdown__swatch"
               :style="selectedSwatchStyle"
               aria-hidden="true"
             />
-            <span class="ml-hatch-pattern-dropdown__label">{{ currentLabel }}</span>
-          </span>
-        </button>
-        <button
-          type="button"
-          class="ml-hatch-pattern-dropdown__caret-button"
-          :disabled="disabled"
-          aria-label="Open hatch pattern list"
-        >
-          <el-icon>
-            <arrow-down />
-          </el-icon>
-        </button>
-      </div>
-
-      <template #dropdown>
-        <ml-hatch-pattern-panel
-          :model-value="normalizedModelValue"
-          :options="resolvedOptions"
-          :disabled="disabled"
-          @select="handlePatternSelect"
-        />
+          </template>
+        </MlRibbonButton>
       </template>
-    </el-dropdown>
+
+      <MlHatchPatternPanel
+        :model-value="normalizedModelValue"
+        :options="resolvedOptions"
+        :disabled="disabled"
+        @select="handlePatternSelect"
+      />
+    </ElPopover>
   </div>
 </template>
 
 <script setup lang="ts">
-import { ArrowDown } from '@element-plus/icons-vue'
-import { computed, ref } from 'vue'
+import { MlRibbonButton } from '@mlightcad/ribbon'
+import { ElPopover } from 'element-plus'
+import { computed, ref, watch } from 'vue'
 
 import {
   DEFAULT_HATCH_PATTERN_OPTIONS,
@@ -66,19 +54,22 @@ import MlHatchPatternPanel from './MlHatchPatternPanel.vue'
  * Props accepted by the hatch pattern dropdown button.
  */
 interface HatchPatternDropdownProps {
+  /** Optional stable id applied to the ribbon button. */
+  id?: string
   /** Current hatch pattern name. */
   modelValue?: string
   /** Candidate hatch patterns available in the picker panel. */
   options?: HatchPatternOption[]
   /** Optional value prefix applied before callback emission. */
   itemIdPrefix?: string
-  /** Disables both trigger buttons and panel interactions. */
+  /** Disables both trigger button and panel interactions. */
   disabled?: boolean
   /** Optional callback when the selected value changes. */
   emitItemClick?: (payload?: string | number | boolean) => void
 }
 
 const props = withDefaults(defineProps<HatchPatternDropdownProps>(), {
+  id: 'hatch-pattern',
   modelValue: 'ANSI31',
   options: () => DEFAULT_HATCH_PATTERN_OPTIONS,
   itemIdPrefix: 'hatch-pattern:',
@@ -92,9 +83,10 @@ const emit = defineEmits<{
 }>()
 
 const isOpen = ref(false)
-const dropdownRef = ref<{ handleClose?: () => void } | null>(null)
 
-const normalizedModelValue = computed(() => props.modelValue.trim().toUpperCase())
+const normalizedModelValue = computed(() =>
+  props.modelValue.trim().toUpperCase()
+)
 
 const resolvedOptions = computed<HatchPatternOption[]>(() =>
   props.options.map(item => ({
@@ -115,6 +107,16 @@ const selectedSwatchStyle = computed(() =>
   resolveHatchPatternSwatchStyle(normalizedModelValue.value)
 )
 
+const buttonId = computed(() => props.id.trim() || 'hatch-pattern')
+const accessibleLabel = computed(() => currentLabel.value || buttonId.value)
+
+watch(
+  () => props.disabled,
+  value => {
+    if (value) isOpen.value = false
+  }
+)
+
 /**
  * Emits change payload for the selected hatch pattern.
  *
@@ -128,14 +130,6 @@ function emitSelection(patternName: string) {
 }
 
 /**
- * Re-applies the current pattern when the primary split-button area is clicked.
- */
-function handleMainButtonClick() {
-  if (props.disabled) return
-  emitSelection(normalizedModelValue.value)
-}
-
-/**
  * Handles panel item selection and closes the dropdown popper.
  *
  * @param patternName Hatch pattern selected in the panel.
@@ -144,105 +138,41 @@ function handlePatternSelect(patternName: string) {
   if (props.disabled) return
   emitSelection(patternName)
   isOpen.value = false
-  dropdownRef.value?.handleClose?.()
 }
 </script>
 
 <style scoped>
 .ml-hatch-pattern-dropdown {
-  display: flex;
+  --ml-hatch-pattern-scale: var(--ml-rb-scale, 1);
+  display: inline-flex;
   width: 100%;
   min-width: 152px;
-  min-height: var(--el-component-size-small, 28px);
 }
 
-.ml-hatch-pattern-dropdown :deep(.el-dropdown) {
-  width: 100%;
-}
-
-.ml-hatch-pattern-dropdown__trigger {
-  display: flex;
-  width: 100%;
-  min-height: var(--el-component-size-small, 28px);
-  border: 1px solid var(--el-border-color);
-  border-radius: var(--el-border-radius-base);
-  background: var(--el-fill-color-blank);
-  overflow: hidden;
-}
-
-.ml-hatch-pattern-dropdown__trigger:hover {
-  border-color: var(--el-border-color-hover);
-}
-
-.ml-hatch-pattern-dropdown__trigger.is-disabled {
-  border-color: var(--el-border-color-light);
-  background: var(--el-fill-color-light);
-}
-
-.ml-hatch-pattern-dropdown__main-button,
-.ml-hatch-pattern-dropdown__caret-button {
-  border: none;
-  background: transparent;
-  color: var(--el-text-color-regular);
-  min-height: var(--el-component-size-small, 28px);
-  cursor: pointer;
-}
-
-.ml-hatch-pattern-dropdown__main-button {
+.ml-hatch-pattern-dropdown :deep(.el-popover__reference-wrapper) {
   display: inline-flex;
-  align-items: center;
-  justify-content: flex-start;
-  flex: 1 1 auto;
   width: 100%;
-  padding: 0 6px;
+}
+
+.ml-hatch-pattern-dropdown__button {
+  width: 100%;
   min-width: 0;
-}
-
-.ml-hatch-pattern-dropdown__caret-button {
-  flex: 0 0 24px;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  border-left: 1px solid var(--el-border-color-light);
-  font-size: 12px;
-}
-
-.ml-hatch-pattern-dropdown__main-button:hover,
-.ml-hatch-pattern-dropdown__caret-button:hover {
-  background: var(--el-fill-color);
-}
-
-.ml-hatch-pattern-dropdown__main-button:disabled,
-.ml-hatch-pattern-dropdown__caret-button:disabled {
-  color: var(--el-disabled-text-color);
-  cursor: not-allowed;
-}
-
-.ml-hatch-pattern-dropdown__value {
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  min-width: 0;
-  width: 100%;
 }
 
 .ml-hatch-pattern-dropdown__swatch {
-  width: 18px;
-  height: 18px;
+  display: inline-block;
+  width: calc(18px * var(--ml-hatch-pattern-scale));
+  height: calc(18px * var(--ml-hatch-pattern-scale));
   box-sizing: border-box;
-  border: 1px solid #404a59;
-  flex: 0 0 18px;
-}
-
-.ml-hatch-pattern-dropdown__label {
-  font-size: 12px;
-  line-height: 1;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
+  border: 1px solid var(--ml-rb-border-strong, #404a59);
+  flex: 0 0 calc(18px * var(--ml-hatch-pattern-scale));
 }
 
 :global(.ml-hatch-pattern-dropdown-popper) {
   padding: 0;
+}
+
+:global(.ml-hatch-pattern-dropdown-popper .el-popper__arrow) {
+  display: none;
 }
 </style>
