+++ packages/cad-viewer/src/component/common/index.ts
@@ -13,10 +13,5 @@ export { default as MlLayerSelect } from './MlLayerSelect.vue'
 export { default as MlFileReader } from './MlFileReader.vue'
 export { default as MlLineTypeSelect } from './MlLineTypeSelect.vue'
 export { default as MlHatchPatternDropdown } from './MlHatchPatternDropdown.vue'
-export { default as MlRibbonHatchPatternButton } from './MlRibbonHatchPatternButton.vue'
-export { default as MlRibbonPropertyColorDropdown } from './MlRibbonPropertyColorDropdown.vue'
-export { default as MlRibbonPropertyField } from './MlRibbonPropertyField.vue'
-export { default as MlRibbonPropertyLineTypeSelect } from './MlRibbonPropertyLineTypeSelect.vue'
-export { default as MlRibbonPropertyLineWeightSelect } from './MlRibbonPropertyLineWeightSelect.vue'
 export { default as MlSysVarToggleButton } from './MlSysVarToggleButton.vue'
 export { default as MlToggleButton } from './MlToggleButton.vue'
