+++ packages/cad-viewer/src/component/index.ts
@@ -2,6 +2,7 @@ export * from './common'
 export * from './dialog'
 export * from './palette'
 export * from './layout'
+export * from './ribbon'
 export * from './notification'
 export * from './statusBar'
 export { default as MlCadViewer } from './MlCadViewer.vue'
