+++ packages/cad-viewer/src/component/layout/index.ts
@@ -3,6 +3,4 @@ export { default as MlEntityInfo } from './MlEntityInfo.vue'
 export { default as MlLanguageSelector } from './MlLanguageSelector.vue'
 export { default as MlLayerDrawStyleToolbar } from './MlLayerDrawStyleToolbar.vue'
 export { default as MlMainMenu } from './MlMainMenu.vue'
-export { default as MlRibbonCommands } from './MlRibbonCommands.vue'
-export { default as MlRibbonLanguageSelector } from './MlRibbonLanguageSelector.vue'
 export { default as MlToolBars } from './MlToolBars.vue'
