+++ packages/cad-viewer/src/composable/index.ts
@@ -17,4 +17,5 @@ export * from './useSettings'
 export * from './useSystemVars'
 export * from './useIsMobile'
 export * from './useQuickSelect'
+export * from './useRibbonContextualTab'
 export * from './useViewerRect'
